refactor: extract file preview functionality into separate components
Split the large files-section.tsx into modular, reusable components:
- Created file-preview-dialog.tsx for the preview UI
- Created json-tree.tsx for JSON data visualization
- Created use-file-preview.ts hook for preview logic
- Created file-preview.ts utilities for file type detection

Reduces main file by ~460 lines while maintaining all functionality
